Move Prometheus metrics initialization to separate class to fix conflict in pulsar-proxy tests
